Added slice, sim-starity, filter-slices, and slice-n-rename
sim-starity works around lack of foresight in certain function
definitions not providing a star-arity version. It's not quite
map.

This is a family of function that perform potentially deep map
walking and return the requested value and the path to that value.
slice is analagous to get', but returns a map containing the entire
path to the value with the value. As such, it "slices" a piece of
a map. filter-slices is like select-keys, but uses get' to support
arbitrary lookups. slice-n-rename takes a map and then arg seqs
containing a from keyspec and a to keyspec. All values found in
from keyspecs will be sliced, merged, and then renames performed
as requested.